# docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_efangly_thanes-lims-backend_internal_adapters_http_response.Envelope:
    properties:
      data: {}
      error:
        $ref: '#/definitions/github_com_efangly_thanes-lims-backend_internal_adapters_http_response.ErrorBody'
      meta: {}
      success:
        type: boolean
    type: object
  github_com_efangly_thanes-lims-backend_internal_adapters_http_response.ErrorBody:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
...
  internal_adapters_http_location.CreateCabinetRequest:
    properties:
      kind:
        description: |-
          Kind selects the tree: "sample_storage" (default) or
          "equipment_storage".
        enum:
        - sample_storage
        - equipment_storage
        type: string
      name:
        type: string
    required:
    - name
    type: object
  internal_adapters_http_location.FullPathResponse:
    properties:
      full_path:
        type: string
    type: object
  internal_adapters_http_location.GenerateChildrenRequest:
    properties:
      count:
        minimum: 1
        type: integer
      prefix:
        type: string
    required:
    - count
    - prefix
    type: object
  internal_adapters_http_location.LocationResponse:
    properties:
      barcode_code:
        type: string
      id:
        type: string
      kind:
        type: string
      level_type:
        type: string
      name:
        type: string
      parent_id:
        type: string
    type: object
...
info:
  contact: {}
  description: REST API สำหรับ Thanes LIMS backend (Auth, Sample/Chain-of-Custody,
    Test Result, Equipment, Inventory/Purchase Order, Document, Environment, Notification,
    Audit)
  title: Thanes LIMS Backend API
  version: "1.0"
paths:
...
  /locations:
    get:
      description: ไม่ระบุ query param `parent_id` จะได้รายการตู้ (root) ทั้งหมด
      parameters:
      - description: Parent Location ID
        in: query
        name: parent_id
        type: string
      - description: Tree kind for root listing (sample_storage|equipment_storage,
          default sample_storage)
        in: query
        name: kind
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_efangly_thanes-lims-backend_internal_adapters_http_response.Envelope'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/internal_adapters_http_location.LocationResponse'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_efangly_thanes-lims-backend_internal_adapters_http_response.Envelope'
      security:
      - BearerAuth: []
      summary: รายการ Location ลูกโดยตรง
      tags:
      - locations
    post:
      consumes:
      - application/json
      description: สร้าง root Location - ชื่อตู้ต้อง unique ทั้งระบบ
      parameters:
      - description: ชื่อตู้
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_adapters_http_location.CreateCabinetRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/github_com_efangly_thanes-lims-backend_internal_adapters_http_response.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/internal_adapters_http_location.LocationResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_efangly_thanes-lims-backend_internal_adapters_http_response.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_efangly_thanes-lims-backend_internal_adapters_http_response.Envelope'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_efangly_thanes-lims-backend_internal_adapters_http_response.Envelope'
      security:
      - BearerAuth: []
      summary: สร้างตู้ (Cabinet) ใหม่
      tags:
      - locations
  /locations/{id}:
    delete:
      description: ลบไม่ได้ถ้ายังมีลูกหรือมี sample อ้างอิงอยู่ (restrict)
      parameters:
      - description: Location ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: no content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_efangly_thanes-lims-backend_internal_adapters_http_response.Envelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_efangly_thanes-lims-backend_internal_adapters_http_response.Envelope'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_efangly_thanes-lims-backend_internal_adapters_http_response.Envelope'
      security:
      - BearerAuth: []
      summary: ลบ Location
      tags:
      - locations
  /locations/{id}/children:
    post:
      consumes:
      - application/json
      description: สร้าง Location ลูกของ :id จำนวน `count` ตัว ชื่อ "{prefix}-1".."{prefix}-{count}"
        ที่ระดับถัดไปในลำดับชั้น (cabinet > shelf > slot > sub_slot)
      parameters:
      - description: Parent Location ID
        in: path
        name: id
        required: true
        type: string
      - description: prefix + จำนวน
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_adapters_http_location.GenerateChildrenRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/github_com_efangly_thanes-lims-backend_internal_adapters_http_response.Envelope'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/internal_adapters_http_location.LocationResponse'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_efangly_thanes-lims-backend_internal_adapters_http_response.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_efangly_thanes-lims-backend_internal_adapters_http_response.Envelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_efangly_thanes-lims-backend_internal_adapters_http_response.Envelope'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_efangly_thanes-lims-backend_internal_adapters_http_response.Envelope'
      security:
      - BearerAuth: []
      summary: สร้างลูกอัตโนมัติตามจำนวนที่ระบุ
      tags:
      - locations
  /locations/{id}/full-path:
    get:
      description: คำนวณจาก tree ทุกครั้งที่เรียก เช่น "Fridge-A / Shelf-2 / Slot-4"
      parameters:
      - description: Location ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_efangly_thanes-lims-backend_internal_adapters_http_response.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/internal_adapters_http_location.FullPathResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_efangly_thanes-lims-backend_internal_adapters_http_response.Envelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_efangly_thanes-lims-backend_internal_adapters_http_response.Envelope'
      security:
      - BearerAuth: []
      summary: รหัสเต็มของ Location (Full Path)
      tags:
      - locations
  /locations/by-barcode/{code}:
    get:
      description: สแกน Location Barcode แล้ว resolve เป็น Location โดยตรง (ใช้ตอนย้าย
        Sample/วางของ)
      parameters:
      - description: Location Barcode (เช่น LOC-BC-00001)
        in: path
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_efangly_thanes-lims-backend_internal_adapters_http_response.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/internal_adapters_http_location.LocationResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_efangly_thanes-lims-backend_internal_adapters_http_response.Envelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_efangly_thanes-lims-backend_internal_adapters_http_response.Envelope'
      security:
      - BearerAuth: []
      summary: ค้นหา Location จาก Barcode
      tags:
      - locations
...
securityDefinitions:
  BearerAuth:
    description: พิมพ์ "Bearer" ตามด้วยเว้นวรรคและ JWT access token เช่น "Bearer eyJhbGciOi..."
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"